backend/requirements.txt
Flask==2.3.2 
Flask-Cors==4.0.0 
bcrypt==4.0.1 
PyJWT==2.8.0